+++ components/ui/index.ts
@@ -1,6 +1,7 @@
 export * from "./Button";
 export * from "./Input";
 export * from "./Textarea";
+export * from "./Select";
 export * from "./Card";
 export * from "./PageHeader";
 export * from "./EmptyState";