+++ features/brands/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./BrandCard";
+export * from "./BrandList";
+export * from "./BrandForm";